re-organiser le code au niveau des pages et des formulaires pour separer les composants qui sont en use client et ceux en server
